@@ -3,27 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.2.2",
+    "@ant-design/icons": "^4.6.2",
+    "@craco/craco": "^6.1.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "antd": "^4.7.2",
+    "antd": "^4.14.1",
+    "copy-to-clipboard": "^3.3.1",
+    "craco-less": "^1.17.1",
     "i18next": "^19.8.9",
     "moment": "^2.29.1",
-    "react": "^16.14.0",
+    "react": "^17.0.2",
     "react-cropper": "^2.1.7",
     "react-device-detect": "^1.14.0",
-    "react-dom": "^16.14.0",
+    "react-dom": "^17.0.2",
     "react-github-corner": "^2.5.0",
     "react-i18next": "^11.8.7",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3"
+    "react-scripts": "4.0.3"
   },
   "scripts": {
-    "start": "set PORT=7001 && react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "set PORT=7001 && craco start",
+    "build": "craco build",
+    "test": "craco test",
+    "eject": "craco eject"
   },
   "eslintConfig": {
     "extends": "react-app"
